Implement JavaScript console :^)
Thanks to Tim's awesome work, we can reuse most of the ConsoleClient logic.
The downside is that the console output is displayed in a WebView. We'll see
if this is something that could be improved upon in the future.

The JS console is currently only accessible through the F12 shortcut.